backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
duckdb==1.1.0
pydantic==2.9.0
python-multipart==0.0.9
google-generativeai>=0.8.0
python-dotenv>=1.0.0
Pillow>=10.0.0
